requirements.txt
Flask==3.0.3
Werkzeug>=3.0.3
cryptography>=42.0.5
requests>=2.31.0
python-dotenv>=1.0.0
